+++ google/cloud/netapp/quickstart/quickstart.cc
@@ -13,21 +13,21 @@
 // limitations under the License.
 
 //! [all]
-#include "google/cloud/netapp/ EDIT HERE .h"
-#include "google/cloud/project.h"
+#include "google/cloud/netapp/v1/net_app_client.h"
 #include <iostream>
 
 int main(int argc, char* argv[]) try {
-  if (argc != 2) {
-    std::cerr << "Usage: " << argv[0] << " project-id\n";
+  if (argc != 3) {
+    std::cerr << "Usage: " << argv[0] << " project-id location-id\n";
     return 1;
   }
 
-  namespace netapp = ::google::cloud::netapp;
-  auto client = netapp::Client(netapp::MakeConnection());
+  namespace netapp = ::google::cloud::netapp_v1;
+  auto client = netapp::NetAppClient(netapp::MakeNetAppConnection());
 
-  auto const project = google::cloud::Project(argv[1]);
-  for (auto r : client.List /*EDIT HERE*/ (project.FullName())) {
+  auto const parent =
+      std::string{"projects/"} + argv[1] + "/locations/" + argv[2];
+  for (auto r : client.ListStoragePools(parent)) {
     if (!r) throw std::move(r).status();
     std::cout << r->DebugString() << "\n";
   }