+++ CHANGELOG.md
@@ -116,6 +116,7 @@ We are happy to announce the following GA libraries. Unless specifically noted,
 the APIs in these libraries are stable, and are ready for production use.
 
 - [Infrastructure Manager](/google/cloud/config/README.md)
+- [NetApp](/google/cloud/config/README.md)
 
 ### [Discovery Engine](/google/cloud/discoveryengine/README.md)